services/ml/requirements.txt
# ML Service Dependencies

# Core ML libraries
sentence-transformers==2.2.2
torch>=2.0.0
transformers>=4.30.0

# Image processing
Pillow>=10.0.0
imagehash>=4.3.1

# API framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.0.0

# Data processing
numpy>=1.24.0
scipy>=1.10.0

# HTTP requests
requests>=2.31.0
aiohttp>=3.9.0

# Database (if needed)
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# Utilities
python-dotenv>=1.0.0
